clockidup login now properly works when no config file exists
The bug was due to the use of the same http.DefaultClient in two calls.
The first call would "wrap" the transport in http.DefaultClient a first
time, and the second call would wrap a second time.

The use of http.DefaultClient should be deprecated I think. Also,
http.DefaultTransport should have some creation function to avoid state
sharing...